requirements.txt
# -----------------------------
# functions/requirements.txt
# -----------------------------
fastapi
uvicorn
requests
python-multipart
torch
torchvision
pillow
opencv-python
ultralytics